Script to Start leap from remote computer

+++ evals/utils/kill-cloud.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
-for line in $(cat cloud-ip);do
+if [ "$#" -ne 1 ]; then
+    echo "Illegal number of parameters"
+    echo "usage:"
+    echo "[ip_file_path]"
+    exit
+fi
+
+filepath=$1
+
+for line in $(cat $filepath);do
 	hostname=`echo $line | grep -E -o '[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}'`
 	echo "killing" $hostname
 	ssh stolet@$hostname 'pkill cloudalgo_main coordinator-main'
